DTFS2-4361 - split out mapSubmittedDeal in deal.submit controller
