+++ trade-finance-manager-api/src/v1/controllers/deal.submit.controller.js
@@ -21,6 +21,7 @@ const dealController = require('./deal.controller');
 const { shouldUpdateDealFromMIAtoMIN } = require('./should-update-deal-from-MIA-to-MIN');
 const { updatePortalDealFromMIAtoMIN } = require('./update-portal-deal-from-MIA-to-MIN');
 const { sendDealSubmitEmails, sendAinMinIssuedFacilitiesAcknowledgement } = require('./send-deal-submit-emails');
+const { mapSubmittedDeal } = require('../mappings/map-submitted-deal');
 
 const getDeal = async (dealId, dealType) => {
   let deal;
@@ -39,155 +40,6 @@ const getDeal = async (dealId, dealType) => {
   return deal;
 };
 
-// const mapDeal = (deal) => {
-//   let mappedDeal;
-
-//   const { dealType } = deal.dealSnapshot;
-
-//   if (dealType === CONSTANTS.DEALS.DEAL_TYPE.GEF) {
-//     mappedDeal = mapGefDealSnapshot(deal);
-
-//     // temporarily return false for dev.
-//     return false;
-//   }
-
-//   if (dealType === CONSTANTS.DEALS.DEAL_TYPE.BSS_EWCS) {
-//     mappedDeal = mapDealSnapshot(deal);
-//   }
-
-//   return mappedDeal;
-// };
-
-const mapGefDeal = (deal) => {
-  const { dealSnapshot, tfm } = deal;
-
-  const mapped = {
-    _id: dealSnapshot._id,
-    dealType: dealSnapshot.dealType,
-    bankSupplyContractID: dealSnapshot.bankInternalRefName,
-    bankAdditionalReferenceName: dealSnapshot.additionalRefName,
-    submissionCount: dealSnapshot.submissionCount,
-    submissionType: dealSnapshot.submissionType,
-    submissionDate: dealSnapshot.submissionDate,
-    status: dealSnapshot.status,
-    ukefDealId: dealSnapshot.ukefDealId,
-    exporter: {
-      companyName: dealSnapshot.exporter.companyName,
-      companiesHouseRegistrationNumber: dealSnapshot.exporter.companiesHouseRegistrationNumber,
-    },
-    // dealCurrency
-    // dealValue
-    tfm,
-  };
-
-  return mapped;
-};
-
-
-const mapBssEwcsDeal = (deal) => {
-  const { dealSnapshot, tfm } = deal;
-
-  const mapped = {
-    _id: dealSnapshot._id,
-    dealType: dealSnapshot.dealType,
-    bankReferenceNumber: dealSnapshot.details.bankSupplyContractID,
-    bankAdditionalReferenceName: dealSnapshot.details.bankSupplyContractName,
-    submissionCount: dealSnapshot.details.submissionCount,
-    submissionType: dealSnapshot.details.submissionType,
-    submissionDate: dealSnapshot.details.submissionDate,
-    status: dealSnapshot.details.status,
-    ukefDealId: dealSnapshot.details.ukefDealId,
-    maker: dealSnapshot.details.maker,
-    exporter: {
-      companyName: dealSnapshot.submissionDetails['supplier-name'],
-      companiesHouseRegistrationNumber: dealSnapshot.submissionDetails['supplier-companies-house-registration-number'],
-    },
-    buyer: {
-      name: dealSnapshot.submissionDetails['buyer-name'],
-      country: dealSnapshot.submissionDetails['buyer-address-country'],
-    },
-    indemnifier: {
-      name: dealSnapshot.submissionDetails['indemnifier-name'],
-    },
-    dealCurrency: dealSnapshot.submissionDetails.supplyContractCurrency,
-    dealValue: dealSnapshot.submissionDetails.supplyContractValue,
-    destinationOfGoodsAndServices: dealSnapshot.submissionDetails.destinationOfGoodsAndServices,
-    eligibility: dealSnapshot.eligibility,
-    // TODO: align BSS/EWCS and GEF facilities.
-    // update functions that consume this data
-    facilities: [
-      ...dealSnapshot.bondTransactions.items.map((facility) => ({
-        _id: facility._id,
-        ukefFacilityID: facility.ukefFacilityID,
-        facilityType: facility.facilityType,
-        coverStartDate: facility.requestedCoverStartDate,
-        'coverEndDate-day': facility['coverEndDate-day'],
-        'coverEndDate-month': facility['coverEndDate-month'],
-        'coverEndDate-year': facility['coverEndDate-year'],
-        ukefGuaranteeInMonths: facility.ukefGuaranteeInMonths,
-        facilityStage: facility.facilityStage,
-        previousFacilityStage: facility.previousFacilityStage,
-        currency: facility.currency,
-        facilityValue: facility.facilityValue,
-        coveredPercentage: facility.coveredPercentage,
-        ukefExposure: facility.ukefExposure,
-        disbursementAmount: facility.disbursementAmount,
-        guaranteeFeePayableByBank: facility.guaranteeFeePayableByBank,
-        dayCountBasis: facility.dayCountBasis,
-        feeFrequency: facility.feeFrequency,
-        feeType: facility.feeType,
-        uniqueIdentificationNumber: facility.uniqueIdentificationNumber,
-        tfm: facility.tfm,
-      })),
-      ...dealSnapshot.loanTransactions.items.map((facility) => ({
-        _id: facility._id,
-        ukefFacilityID: facility.ukefFacilityID,
-        facilityType: facility.facilityType,
-        coverStartDate: facility.requestedCoverStartDate,
-        'coverEndDate-day': facility['coverEndDate-day'],
-        'coverEndDate-month': facility['coverEndDate-month'],
-        'coverEndDate-year': facility['coverEndDate-year'],
-        ukefGuaranteeInMonths: facility.ukefGuaranteeInMonths,
-        facilityStage: facility.facilityStage,
-        previousFacilityStage: facility.previousFacilityStage,
-        currency: facility.currency,
-        facilityValue: facility.facilityValue,
-        coveredPercentage: facility.coveredPercentage,
-        ukefExposure: facility.ukefExposure,
-        disbursementAmount: facility.disbursementAmount,
-        guaranteeFeePayableByBank: facility.guaranteeFeePayableByBank,
-        dayCountBasis: facility.dayCountBasis,
-        feeFrequency: facility.feeFrequency,
-        feeType: facility.feeType,
-        uniqueIdentificationNumber: facility.uniqueIdentificationNumber,
-        tfm: facility.tfm,
-      })),
-    ],
-    tfm,
-  };
-
-  return mapped;
-};
-
-const unifiedDealStructure = (deal) => {
-  let mappedDeal;
-
-  const { dealType } = deal.dealSnapshot;
-
-  if (dealType === CONSTANTS.DEALS.DEAL_TYPE.GEF) {
-    mappedDeal = mapGefDeal(deal);
-
-    // temporarily return false for dev.
-    return false;
-  }
-
-  if (dealType === CONSTANTS.DEALS.DEAL_TYPE.BSS_EWCS) {
-    mappedDeal = mapBssEwcsDeal(deal);
-  }
-
-  return mappedDeal;
-};
-
 const submitDeal = async (dealId, dealType, checker) => {
   const deal = await getDeal(dealId, dealType);
 
@@ -195,14 +47,9 @@ const submitDeal = async (dealId, dealType, checker) => {
     return false;
   }
 
-  // TODO:
-  // 1) use mapDeal (and maybe dealReducer functionality??) to bring both deal types into nice data shape.
-  // 2) update variable names in functions below. Simpler, cleaner. One step closer to unified structure.
-  // 3) extract 'mappings' so it's outside of graphql as it won't be GQL specific
-
   const submittedDeal = await api.submitDeal(dealId);
 
-  const mappedDeal = unifiedDealStructure(submittedDeal);
+  const mappedDeal = mapSubmittedDeal(submittedDeal);
 
   const { submissionCount } = mappedDeal;
